requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.11.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
loguru==0.7.0
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
text-unidecode==1.3
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
